new testing module and tests refactor
