+++ doc/whats-new.rst
@@ -168,9 +168,9 @@ Enhancements
   :py:class:`FacetGrid` and :py:func:`~xarray.plot.plot`, so axes
   sharing can be disabled for polar plots.
   By `Bas Hoonhout <https://github.com/hoonhout>`_.
-- New utility functions :py:func:`~xarray.test.assert_xarray_equal`,
-  :py:func:`~xarray.test.assert_xarray_identical`, and
-  :py:func:`~xarray.test.assert_xarray_allclose` for asserting relationships
+- New utility functions :py:func:`~xarray.testing.assert_equal`,
+  :py:func:`~xarray.testing.assert_identical`, and
+  :py:func:`~xarray.testing.assert_allclose` for asserting relationships
   between xarray objects, designed for use in a pytest test suite.
 - ``figsize``, ``size`` and ``aspect`` plot arguments are now supported for all
   plots (:issue:`897`). See :ref:`plotting.figsize` for more details.
