+++ xarray/tests/__init__.py
@@ -12,6 +12,9 @@ import pytest
 from xarray.core import utils, nputils, ops
 from xarray.core.variable import as_variable
 from xarray.core.pycompat import PY3
+from xarray.testing import (decode_string_data, data_allclose_or_equiv,
+                            assert_dataset_allclose, assert_equal,
+                            assert_identical, assert_allclose)
 
 try:
     import unittest2 as unittest
@@ -113,35 +116,6 @@ def requires_bottleneck(test):
     return test if has_bottleneck else pytest.mark.skip('requires bottleneck')(test)
 
 
-def decode_string_data(data):
-    if data.dtype.kind == 'S':
-        return np.core.defchararray.decode(data, 'utf-8', 'replace')
-    return data
-
-
-def data_allclose_or_equiv(arr1, arr2, rtol=1e-05, atol=1e-08):
-    if any(arr.dtype.kind == 'S' for arr in [arr1, arr2]):
-        arr1 = decode_string_data(arr1)
-        arr2 = decode_string_data(arr2)
-    exact_dtypes = ['M', 'm', 'O', 'U']
-    if any(arr.dtype.kind in exact_dtypes for arr in [arr1, arr2]):
-        return ops.array_equiv(arr1, arr2)
-    else:
-        return ops.allclose_or_equiv(arr1, arr2, rtol=rtol, atol=atol)
-
-
-def assert_dataset_allclose(d1, d2, rtol=1e-05, atol=1e-08):
-    assert sorted(d1, key=str) == sorted(d2, key=str)
-    assert sorted(d1.coords, key=str) == sorted(d2.coords, key=str)
-    for k in d1:
-        v1 = d1.variables[k]
-        v2 = d2.variables[k]
-        assert v1.dims == v2.dims
-        allclose = data_allclose_or_equiv(
-            v1.values, v2.values, rtol=rtol, atol=atol)
-        assert allclose, (k, v1.values, v2.values)
-
-
 class TestCase(unittest.TestCase):
     if PY3:
         # Python 3 assertCountEqual is roughly equivalent to Python 2
@@ -158,13 +132,13 @@ class TestCase(unittest.TestCase):
             assert any(message in str(wi.message) for wi in w)
 
     def assertVariableEqual(self, v1, v2):
-        assert_xarray_equal(v1, v2)
+        assert_equal(v1, v2)
 
     def assertVariableIdentical(self, v1, v2):
-        assert_xarray_identical(v1, v2)
+        assert_identical(v1, v2)
 
     def assertVariableAllClose(self, v1, v2, rtol=1e-05, atol=1e-08):
-        assert_xarray_allclose(v1, v2, rtol=rtol, atol=atol)
+        assert_allclose(v1, v2, rtol=rtol, atol=atol)
 
     def assertVariableNotEqual(self, v1, v2):
         assert not v1.equals(v2)
@@ -176,77 +150,25 @@ class TestCase(unittest.TestCase):
         assert a1 == a2 or (a1 != a1 and a2 != a2)
 
     def assertDatasetEqual(self, d1, d2):
-        assert_xarray_equal(d1, d2)
+        assert_equal(d1, d2)
 
     def assertDatasetIdentical(self, d1, d2):
-        assert_xarray_identical(d1, d2)
+        assert_identical(d1, d2)
 
     def assertDatasetAllClose(self, d1, d2, rtol=1e-05, atol=1e-08):
-        assert_xarray_allclose(d1, d2, rtol=rtol, atol=atol)
+        assert_allclose(d1, d2, rtol=rtol, atol=atol)
 
     def assertCoordinatesEqual(self, d1, d2):
-        assert_xarray_equal(d1, d2)
+        assert_equal(d1, d2)
 
     def assertDataArrayEqual(self, ar1, ar2):
-        assert_xarray_equal(ar1, ar2)
+        assert_equal(ar1, ar2)
 
     def assertDataArrayIdentical(self, ar1, ar2):
-        assert_xarray_identical(ar1, ar2)
+        assert_identical(ar1, ar2)
 
     def assertDataArrayAllClose(self, ar1, ar2, rtol=1e-05, atol=1e-08):
-        assert_xarray_allclose(ar1, ar2, rtol=rtol, atol=atol)
-
-
-def assert_xarray_equal(a, b):
-    import xarray as xr
-    ___tracebackhide__ = True  # noqa: F841
-    assert type(a) == type(b)
-    if isinstance(a, (xr.Variable, xr.DataArray, xr.Dataset)):
-        assert a.equals(b), '{}\n{}'.format(a, b)
-    else:
-        raise TypeError('{} not supported by assertion comparison'
-                        .format(type(a)))
-
-
-def assert_xarray_identical(a, b):
-    import xarray as xr
-    ___tracebackhide__ = True  # noqa: F841
-    assert type(a) == type(b)
-    if isinstance(a, xr.DataArray):
-        assert a.name == b.name
-        assert_xarray_identical(a._to_temp_dataset(), b._to_temp_dataset())
-    elif isinstance(a, (xr.Dataset, xr.Variable)):
-        assert a.identical(b), '{}\n{}'.format(a, b)
-    else:
-        raise TypeError('{} not supported by assertion comparison'
-                        .format(type(a)))
-
-
-def assert_xarray_allclose(a, b, rtol=1e-05, atol=1e-08):
-    import xarray as xr
-    ___tracebackhide__ = True  # noqa: F841
-    assert type(a) == type(b)
-    if isinstance(a, xr.Variable):
-        assert a.dims == b.dims
-        allclose = data_allclose_or_equiv(
-            a.values, b.values, rtol=rtol, atol=atol)
-        assert allclose, '{}\n{}'.format(a.values, b.values)
-    elif isinstance(a, xr.DataArray):
-        assert_xarray_allclose(a.variable, b.variable)
-        for v in a.coords.variables:
-            # can't recurse with this function as coord is sometimes a DataArray,
-            # so call into data_allclose_or_equiv directly
-            allclose = data_allclose_or_equiv(
-                a.coords[v].values, b.coords[v].values, rtol=rtol, atol=atol)
-            assert allclose, '{}\n{}'.format(a.coords[v].values, b.coords[v].values)
-    elif isinstance(a, xr.Dataset):
-        assert sorted(a, key=str) == sorted(b, key=str)
-        for k in list(a.variables) + list(a.coords):
-            assert_xarray_allclose(a[k], b[k], rtol=rtol, atol=atol)
-
-    else:
-        raise TypeError('{} not supported by assertion comparison'
-                        .format(type(a)))
+        assert_allclose(ar1, ar2, rtol=rtol, atol=atol)
 
 
 class UnexpectedDataAccess(Exception):
