+++ xarray/tests/data/bears.nc
