+++ xarray/tests/test_combine.py
