+++ xarray/tests/test_groupby.py
