+++ xarray/tests/test_indexing.py
