+++ xarray/tests/test_plot.py
